mcp-repo-connector/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
aiofiles>=23.2.1
python-multipart>=0.0.9
mcp>=0.1.0
httpx>=0.26.0
jinja2>=3.1.3
